OSLinux: Use XDG_DATA_HOME for user data dir
So far it was even set to $HOME/Library which I think is not a very good
default and looks a lot like it was copied over from the OS X
implementation.

The XDG Base Directory Specification[1] is out there since about 10
years or even more, so let's look up the XDG_DATA_HOME environment
variable and if it's not set, fall back to ~/.local/share as in the
specification.

[1]: https://standards.freedesktop.org/basedir-spec/basedir-spec-latest.html

Signed-off-by: aszlig <aszlig@nix.build>