Fixes a bug when >1 entities are added and removed to/from 1 family
The bug would occur under the following circumstances:
* 2 Entities (1 and 2) belong to one family
* Both of those entities had an optional component added to them in the same frame
And it would lead to only one of the 2 entities to be updated

This lead to the following situation:
toRemove = { 1, 2 }
entities = { 1(old), 2(old), 1(new), 2(new) }
after one pass it would then lead to this situation:
toRemove = { 2 }
entities = { 2(new), 2(old), 1(new), 1(old) }
And on the second pass to:
toRemove = {}
entities = { 1(new), 2(old), 2(new), 1(old) }
Afterwards deleting the wrong entity with the ID 2.

The fix buffers the swaps in a vector and executes them all at once, preventing the changes from interfering with the forward scan.
Using reference instead of copy in iteration through moves