Temp fix? Editor crash when reparenting entities
Problem is that when children are moved down the hierachy into a new parent, they will be removed first.
After that the engine tries to apply the new parent to that removed child and crashes.
To fix that I buffer the deletion in the context and delete the children at the end of the delta operation.
Is probably not the real solution but fixes it for now